Account for scripts loading on the client
Global Loot Modifiers don't get loaded on the client, and the map is
never initialized. This would cause a crash, but now the exception is
caught and logged.

We could bypass the need via an injected getter and an explicit null
check, but the amount of times this will happen is minuscule compared to
the amount of checks. Basically, the amount of queries is usually way
less than the amount of registrations and removals.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>