Add builders for vanilla-provided loot conditions
Some builders are still missing, those are marked with TODOs on the
code. Some particular care has to be taken in this case due to how
convoluted vanilla makes those, with lots and lots of moving pieces.

These builders represent a way to make building conditions easier for
players, since functions are provided for them already, using a simple
builder pattern. The core of this functionality is
CTLootConditionBuilder, which provides an entry point for all the other
builders, allowing for chaining.

A JSON Loot Condition is also provided so that people can leverage IData
if they want more control (for whatever control JSON provides) or want
to integrate with other mod's loot conditions.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>